Add support for raw values in pixels.
That is, if you say 'rawWidth: 1080', then your window would be set to
1080 wide.

If you have both raw values and the existing units of 12 values, the raw
values are used.

All of this allows very precise window geometry to be specified, at the
noticible cost that any configuration given in pixels is going to be
specific to that monitor resolution.

Also note, the coordinates are given for a given desktop and screen,
instead of in the global address space.

The latter would 'solve' the multi-monitor problem, but that would end
up with an entirely different way of thinking about coordinates between
the raw values and the plain values.

Though I do need to make a PR for the Window Geometry Information kwin
script, as that currently only gives window coordinates in the global
coordinate state.

Here is an example window configuration using the new functionality:

        {
            rawY: 651,
            rawWidth: 2537,
            rawHeight: 1439,
            x: 0.00,
            y: 4.01,
            width: 5.94,
            height: 7.99,
            shortcutModifier: Qt.AltModifier,
            shortcutKey: Qt.Key_1
        },