Clear shortcuts on hide.
We add the shortcuts to the list every time that we bring up the dialog,
which means that we slowly accumulate a decent number of them, and all
of the associated data structures.

In the more extreme cases, this can absolutely end up with us taking out
kwin entirely with references to stale objects.