Rework how we do layouts for the dialog.
There are a few different things here, all of them important.

We are bounding the size of the boxes to 12, though we are not bothering
to ensure that the x/y + width/height comes out to 12 or less, so we may
still draw outside of the box, but not as much at least.

We are using Math.round to do a round-nearest on the resulting
x/y/width/height value, ensuring that we're not only dealing with
integers, but with integers that have been correctly rounded.

And we're passing screen into the WindowLayout object, so that we don't
get errors about binding to a non-NOTIFYable variable.

Oh, and a typo fix.