Reload the configuration when we open the dialog.
This is a giant hack, but it works.

Nothing causes the configChanged signal to be sent when a KWin script
has it's settings changed, which means that even if we try to reload the
configuration settings, we just get the cached values.

But it turns out that we can send the signal ourselves, forcing KWin to
reload all of it's configuration settings.

If we then hook into the signal this way, as opposed to the previous
way, we will be notified when everything has actually been handled
inside KWin, and we can then correctly load any new settings.

The upshot of all of this is: We can now properly load any newly changed
configuration settings.